Оптимизации
Добавлено каскадное удаление, в некоторых запросах доступен доступ всем пользователям, фронтенд загружается из докерхаб

+++ backend/recipes/api.py
@@ -8,9 +8,10 @@ from db import database
 from recipes import schemas
 from recipes.models import Ingredient, Recipe, Tag
 from services import query_list
-from settings import STATIC_ROOT
+from settings import STATIC_ROOT, NOT_FOUND
 from users.models import User
 from users.utils import get_current_user
+from starlette.requests import Request
 
 recipe_router = APIRouter(prefix='/api', tags=["recipe"])
 db_user = User(database)
@@ -19,13 +20,9 @@ db_ingredient = Ingredient(database)
 db_recipe = Recipe(database)
 
 
-@recipe_router.post("/tags/")
-async def create_tag(
-    tag: schemas.Tag,
-    user_dict: User = Depends(get_current_user)
-):
-    if user_dict["is_superuser"] or user_dict["is_staff"] == True:
-        e = await db_tag.create_tag(tag)
+async def __create_tag_ingredient(model, item, user):
+    if user.is_superuser or user.is_staff is True:
+        e = await model(item)
         if type(e) != int:
             e = str(e).split(": ")[-1]
             return JSONResponse(f"Incorrect {e}", status.HTTP_400_BAD_REQUEST)
@@ -33,49 +30,30 @@ async def create_tag(
     return Response(status_code=status.HTTP_403_FORBIDDEN)
 
 
+@recipe_router.post("/tags/")
+async def create_tag(tag: schemas.Tag, user: User = Depends(get_current_user)):
+    return await __create_tag_ingredient(db_tag.create_tag, tag, user)
+
+
 @recipe_router.get("/tags/")
-async def tags():
-    return await db_tag.get_tags()
-
-
 @recipe_router.get("/tags/{pk}/")
-async def tag(pk: int):
-    return (
-        await db_tag.get_tags(pk)
-        or JSONResponse("NotFound", status.HTTP_404_NOT_FOUND)
-    )
+async def tags(pk: int = None):
+    return await db_tag.get_tags(pk) or NOT_FOUND
 
 
 @recipe_router.post("/ingredients/")
 async def create_ingredient(
     ingredient: schemas.Ingredient,
-    user_dict: User = Depends(get_current_user)
+    user: User = Depends(get_current_user)
 ):
-    if user_dict["is_superuser"] or user_dict["is_staff"] == True:
-        e = await db_ingredient.create_ingredient(ingredient)
-        if type(e) != int:
-            e = str(e).split(": ")[-1]
-            return JSONResponse(f"Incorrect {e}", status.HTTP_400_BAD_REQUEST)
-        return Response(status_code=status.HTTP_200_OK)
-    return Response(status_code=status.HTTP_403_FORBIDDEN)
+    return await __create_tag_ingredient(
+        db_ingredient.create_ingredient, ingredient, user)
 
 
 @recipe_router.get("/ingredients/")
-async def ingredients():
-    return await db_ingredient.get_ingredient()
-
-
 @recipe_router.get("/ingredients/{pk}/")
-async def ingredient(pk: int):
-    return (
-        await db_ingredient.get_ingredient(pk)
-        or JSONResponse("NotFound", status.HTTP_404_NOT_FOUND)
-    )
-
-
-@recipe_router.get("/recipes/")
-async def пуе_recipe():
-    return await query_list(await db_recipe.get_recipe_by_id(None))
+async def ingredients(pk: int = None):
+    return await db_ingredient.get_ingredient(pk) or NOT_FOUND
 
 
 @recipe_router.post("/recipes/")
@@ -90,36 +68,43 @@ async def create_recipe(
 ):
 
     filename = image.filename
-    if not filename.lower().endswith(('.jpg', '.jpeg', 'bmp', 'png')):
-        JSONResponse("NotFound", status.HTTP_404_NOT_FOUND)
+    if not filename.lower().endswith(('.jpg', '.jpeg', '.bmp', '.png')):
+        return JSONResponse(
+            {"detail": "Invalid image format"}, status.HTTP_400_BAD_REQUEST)
 
     try:
         async with aiofiles.open(
             os.path.join(STATIC_ROOT, filename), "wb"
         ) as buffer:
             await buffer.write(await image.read())
-        raise
     except Exception:
         buffer = os.path.join(STATIC_ROOT, filename)
         if os.path.isfile(buffer):
             os.remove(buffer)
-        Response({"detail" "NotFound"}, status.HTTP_404_NOT_FOUND)
+        return JSONResponse(
+            {"detail": "Error image"}, status.HTTP_400_BAD_REQUEST)
 
-    print("====", name, text, filename, cooking_time, ingredients, tags)
+    recipe_item = {
+        "author_id": user_dict["id"],
+        "name": name,
+        "text": text,
+        "image": filename,
+        "cooking_time": cooking_time
+    }
     pk = await db_recipe.create_recipe(
-        name, text, filename, cooking_time, ingredients, tags, user_dict["id"]
+        recipe_item, ingredients, tags
     )
     return await db_recipe.get_recipe_by_id(pk)
 
 
+@recipe_router.get("/recipes/")
 @recipe_router.get("/recipes/{pk}/")
-async def recipes(pk: int):
-    recipe_dict = await db_recipe.get_recipe_by_id(pk)
-    if recipe_dict:
-        return recipe_dict[0]
-    return JSONResponse("Страница не найдена.", status.HTTP_404_NOT_FOUND)
-
-
 @recipe_router.delete("/recipes/{pk}/")
-async def delete_recipe(pk: int):
-    return await db_recipe.delete_recipe(pk)
+async def recipes(request: Request, pk: int = None):
+    if request.method == "DELETE":
+        await db_recipe.delete_recipe(pk)
+        return Response(status_code=status.HTTP_204_NO_CONTENT)
+    if pk:
+        recipe_dict = await db_recipe.get_recipe_by_id(pk)
+        return recipe_dict[0] if recipe_dict else NOT_FOUND
+    return await query_list(await db_recipe.get_recipe_by_id(pk))
